backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv>=1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pymongo==4.6.1
motor==3.3.2
redis==5.0.1

# AI/ML
langchain==0.2.0
langchain-community==0.2.0
openai==1.8.0
# sentence-transformers==2.2.2  # Temporarily commented out for faster build
chromadb==0.4.18

# Utils
httpx==0.26.0
pandas==2.1.4
numpy==1.26.3
PyPDF2==3.0.1

# MarkitDown dependencies
markitdown[all]>=0.1.2
aiofiles==23.2.1
sentence-transformers>=2.3.1

# Audio processing
ffmpeg-python>=0.2.0

# Medical format support (optional)
# pydicom>=2.4.0  # For DICOM files

# OCR dependencies for image text extraction (optional)
# paddlepaddle>=2.5.0
# paddleocr>=2.7.0

# Azure Document Intelligence (optional)
# azure-ai-formrecognizer>=3.3.0